Moved some generic classes to modules, added punctuation to tetris.
