+++ JAGEgradle/JAGEmodules/src/main/java/com/valarion/gameengine/eventconditions/InputCondition.java
@@ -1,0 +1,58 @@
+/*******************************************************************************
+ * The MIT License (MIT)
+ * 
+ * Copyright (c) 2015 Rubén Tomás Gracia
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ ******************************************************************************/
+package com.valarion.gameengine.eventconditions;
+
+import org.newdawn.slick.GameContainer;
+import org.newdawn.slick.SlickException;
+import org.w3c.dom.Element;
+
+import com.valarion.gameengine.core.Condition;
+import com.valarion.gameengine.core.Event;
+import com.valarion.gameengine.core.SubTiledMap;
+import com.valarion.gameengine.gamestates.Controls;
+
+/**
+ * Condition of a keypress.
+ * @author Rubén Tomás Gracia
+ *
+ */
+public class InputCondition implements Condition{
+	protected int key = -1;
+
+	@Override
+	public boolean eval(Event e, GameContainer container, SubTiledMap map) {
+		return container.getInput().isKeyPressed(key);
+	}
+
+	@Override
+	public void load(Element node, Object context) throws SlickException {
+		try {
+			key = Controls.class.getDeclaredField(node.getAttribute("key")).getInt(null);
+		} catch (IllegalArgumentException | IllegalAccessException | NoSuchFieldException | SecurityException e) {
+
+			e.printStackTrace();
+		}
+	}
+
+}
