+++ JAGEgradle/JAGEmodules/src/main/java/com/valarion/gameengine/events/rpgmaker/flow/SavePosition.java
@@ -1,0 +1,66 @@
+/*******************************************************************************
+ * The MIT License (MIT)
+ * 
+ * Copyright (c) 2015 Rubén Tomás Gracia
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ ******************************************************************************/
+package com.valarion.gameengine.events.rpgmaker.flow;
+
+import org.newdawn.slick.GameContainer;
+import org.newdawn.slick.SlickException;
+import org.w3c.dom.Element;
+
+import com.valarion.gameengine.core.Event;
+import com.valarion.gameengine.core.SubTiledMap;
+import com.valarion.gameengine.events.rpgmaker.FlowEventClass;
+import com.valarion.gameengine.gamestates.Database;
+
+/**
+ * Event that saves the position of an event in two variables.
+ * @author Rubén Tomás Gracia
+ *
+ */
+public class SavePosition extends FlowEventClass {
+	int xvar = -1, yvar=-1;
+	@Override
+	public void loadEvent(Element node, Object context) throws SlickException {
+		super.loadEvent(node, context);
+		try {
+			xvar = Integer.parseInt(node.getAttribute("xvar"));
+		}
+		catch(Exception e) {}
+		try {
+			yvar = Integer.parseInt(node.getAttribute("yvar"));
+		}
+		catch(Exception e) {}
+	}
+	
+
+	@Override
+	public void performAction(GameContainer container, SubTiledMap map, Event e)
+			throws SlickException {
+		if(xvar >= 0) {
+			Database.instance().getContext().getGlobalVars()[xvar] = getEvent().getXPos();
+		}
+		if(yvar >= 0) {
+			Database.instance().getContext().getGlobalVars()[yvar] = getEvent().getYPos();
+		}
+	}
+}
