+++ JAGEgradle/JAGEmodules/src/main/java/com/valarion/gameengine/events/rpgmaker/flow/Wait.java
@@ -1,0 +1,68 @@
+/*******************************************************************************
+ * The MIT License (MIT)
+ * 
+ * Copyright (c) 2015 Rubén Tomás Gracia
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ ******************************************************************************/
+package com.valarion.gameengine.events.rpgmaker.flow;
+
+import org.newdawn.slick.GameContainer;
+import org.newdawn.slick.SlickException;
+import org.w3c.dom.Element;
+
+import com.valarion.gameengine.core.Event;
+import com.valarion.gameengine.core.SubTiledMap;
+import com.valarion.gameengine.events.rpgmaker.FlowEventClass;
+
+/**
+ * Event that waits a given amount of time.
+ * @author Rubén Tomás Gracia
+ *
+ */
+public class Wait extends FlowEventClass {
+	long waittime, waited;
+	
+	@Override
+	public void loadEvent(Element node, Object context) throws SlickException {
+		super.loadEvent(node, context);
+		
+		waittime = waited = Integer.parseInt(node.getAttribute("time"));
+	}
+	
+	@Override
+	public void performAction(GameContainer container, SubTiledMap map, Event e)
+			throws SlickException {
+		super.performAction(container, map, e);
+		waited = 0;
+	}
+	
+	@Override
+	public void paralelupdate(GameContainer container, int delta,
+			SubTiledMap map) throws SlickException {
+		if(isWorking()) {
+			waited += delta;
+		}
+	}
+	
+	@Override
+	public boolean isWorking() {
+		return waited < waittime;
+	}
+}
